refactoring: moved api and network modules to manager
