--- src/config/__init__.py
@@ -1,15 +1,0 @@
-# SPDX-License-Identifier: MIT
-# --------------------------------------------------------------
-# |																|
-# |             Copyright 2023 - 2023, Amulya Paritosh			|
-# |																|
-# |  This file is part of Component Library Plugin for FreeCAD.	|
-# |																|
-# |               This file was created as a part of				|
-# |              Google Summer Of Code Program - 2023			|
-# |																|
-# --------------------------------------------------------------
-
-from .settings import Settings
-
-config = Settings()
