+++ src/interface/forms/component_upload/upload_form.py
@@ -22,7 +22,7 @@ from ....config import config
 from ..Ui_component_form import Ui_ComponentCreationForm
 
 
-class ComponetUploadDialog(QWidget):
+class ComponentUploadDialog(QWidget):
     """
     Dialog for input fields of data for creating a Component.
     """
