+++ src/manager/api_manager/base_query.py
