--- src/manager/authentication/__init__.py
@@ -1,1 +1,0 @@
-from .authenticator import Authentication_Manager
